+++ src/Labels/Hooks.cpp
@@ -1,12 +1,12 @@
 #include "Hooks.hpp"
-#include "SimpleLabelNode.hpp"
+#include "Nodes/LabelContainerLayer.hpp"
 
 bool LabelPlayLayer::init(GJGameLevel* level, bool useReplay, bool dontCreateObjects)
 {
     if (!PlayLayer::init(level, useReplay, dontCreateObjects))
         return false;
 
-    this->addChild(SimpleLabelNode::create());
+    m_uiLayer->addChild(LabelContainerLayer::create());
 
     return true;
 }