+++ src/Labels/LabelConfig.hpp
@@ -1,0 +1,26 @@
+#pragma once
+
+#include <Geode/Geode.hpp>
+#include "LabelAnchor.hpp"
+
+enum class LabelType
+{
+    Text,
+};
+
+struct LabelConfig
+{
+    std::string displayName = "";
+    std::string formatString = "";
+    float scale = 1.0f;
+    float opacity = 1.0f;
+    std::string font = "bigFont.fnt";
+    LabelAnchor anchor = LabelAnchor::TopLeft;
+    cocos2d::CCPoint offset = cocos2d::CCPointZero;
+    LabelType type = LabelType::Text;
+    bool cheatIndicator = false;
+    bool noclipOnly = false;
+    bool visible = true;
+
+    static LabelConfig createFromObject(matjson::Value obj);
+};