+++ src/Labels/Nodes/LabelNode.hpp
@@ -1,0 +1,25 @@
+#pragma once
+
+#include <Geode/Geode.hpp>
+#include "../LabelConfig.hpp"
+
+using namespace geode::prelude;
+
+class LabelNode : public CCNode
+{
+    protected:
+        LabelConfig config;
+
+        virtual bool init();
+
+    public:
+        void setLabelConfig(LabelConfig config);
+        const LabelConfig& getLabelConfig();
+
+        void updateGeneral(float dt);
+        bool isActionActive();
+
+        virtual void setup();
+        virtual void labelConfigUpdated();
+        virtual void update(float dt);
+};