+++ src/Labels/Nodes/TextLabelNode.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+#include "LabelNode.hpp"
+#include <rift.hpp>
+
+class TextLabelNode : public LabelNode
+{
+    protected:
+        CCLabelBMFont* label = nullptr;
+        rift::Script* script = nullptr;
+
+        ~TextLabelNode();
+
+    public:
+        CREATE_FUNC(TextLabelNode);
+
+        ccColor3B getDesiredColour();
+        void updateVariables();
+
+        virtual void setup();
+        virtual void labelConfigUpdated();
+        virtual void update(float dt);
+};